# docs/swagger/ims-api.yaml
openapi: 3.0.0
info:
  title: IMS Integration API
  description: API for managing IMS (Integration Management System) connections and operations
  version: 1.0.0
  contact:
    name: API Support
    email: support@imsgateway.com

servers:
  - url: https://api.imsgateway.com/v1
    description: Production server
  - url: https://api-test.imsgateway.com/v1
    description: Test server

security:
  - BearerAuth: []

paths:
  /ims-connections:
    get:
      summary: List IMS connections
      description: Retrieves all IMS connections for the authenticated user
      operationId: listConnections
      tags:
        - IMS Connections
      responses:
        '200':
          description: List of IMS connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IMSConnectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create IMS connection
      description: Creates a new IMS connection with the provided credentials
      operationId: createConnection
      tags:
        - IMS Connections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIMSConnectionRequest'
      responses:
        '201':
          description: Connection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IMSConnectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /ims-connections/{connectionId}:
    parameters:
      - name: connectionId
        in: path
        required: true
        schema:
          type: string
        description: ID of the IMS connection

    get:
      summary: Get IMS connection
      description: Retrieves details of a specific IMS connection
      operationId: getConnection
      tags:
        - IMS Connections
      responses:
        '200':
          description: Connection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IMSConnectionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update IMS connection
      description: Updates an existing IMS connection
      operationId: updateConnection
      tags:
        - IMS Connections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIMSConnectionRequest'
      responses:
        '200':
          description: Connection updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IMSConnectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete IMS connection
      description: Deletes an existing IMS connection
      operationId: deleteConnection
      tags:
        - IMS Connections
      responses:
        '204':
          description: Connection deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /ims-connections/{connectionId}/test:
    post:
      summary: Test IMS connection
      description: Tests the connection and validates credentials
      operationId: testConnection
      tags:
        - IMS Connections
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connection test results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestConnectionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    IMSCredentials:
      type: object
      required:
        - username
        - password
        - accountId
        - url
      properties:
        username:
          type: string
          description: IMS account username
        password:
          type: string
          format: password
          description: IMS account password
        accountId:
          type: string
          description: IMS account identifier
        url:
          type: string
          format: uri
          description: IMS API endpoint URL

    CreateIMSConnectionRequest:
      type: object
      required:
        - connectionName
        - environment
        - credentials
      properties:
        connectionName:
          type: string
          maxLength: 100
          description: Name of the connection
        environment:
          type: string
          enum: [Production, Test, Development]
          description: Environment type
        credentials:
          $ref: '#/components/schemas/IMSCredentials'

    UpdateIMSConnectionRequest:
      type: object
      properties:
        connectionName:
          type: string
          maxLength: 100
          description: Updated name of the connection
        credentials:
          $ref: '#/components/schemas/IMSCredentials'
        isActive:
          type: boolean
          description: Connection status flag

    IMSConnectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the connection
        connectionName:
          type: string
          description: Name of the connection
        environment:
          type: string
          enum: [Production, Test, Development]
        accountId:
          type: string
          description: IMS account identifier
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
        status:
          type: string
          enum: [Pending, Active, Failed, Disabled]
        permissions:
          type: object
          additionalProperties:
            type: boolean

    TestConnectionResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the connection test was successful
        message:
          type: string
          description: Details about the test result
        status:
          type: string
          enum: [Pending, Active, Failed, Disabled]
        permissions:
          type: object
          additionalProperties:
            type: boolean

    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
        traceId:
          type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

tags:
  - name: IMS Connections
    description: Operations for managing IMS connections 